requirements.txt
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
Flask==3.0.3
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
sentry-sdk==2.12.0
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2